​

+++ src/main/kotlin/com/zunza/buythedip/external/binance/dto/KlineRestApiResponse.kt
@@ -7,15 +7,15 @@ import java.math.BigDecimal
 @JsonDeserialize(using = KlineDeserializer::class)
 data class KlineRestApiResponse(
     val openTime: Long,
-    val open: BigDecimal,
-    val high: BigDecimal,
-    val low: BigDecimal,
-    val close: BigDecimal,
-    val volume: BigDecimal,
+    val open: Double,
+    val high: Double,
+    val low: Double,
+    val close: Double,
+    val volume: Double,
     val closeTime: Long,
-    val quoteAssetVolume: BigDecimal,
+    val quoteAssetVolume: Double,
     val numberOfTrades: Long,
-    val takerBuyBaseVolume: BigDecimal,
-    val takerBuyQuoteVolume: BigDecimal,
-    val ignore: BigDecimal
+    val takerBuyBaseVolume: Double,
+    val takerBuyQuoteVolume: Double,
+    val ignore: Double
 )
